requirements.txt
# パッケージ更新は `pip3 install -r requirements.txt`で行う
openai
python-dotenv

langchain
langchain-community
langchain-openai

pymupdf
spacy
tiktoken
chromadb

chainlit
wikipedia